+++ gremlin-go/driver/response.go
@@ -23,7 +23,7 @@ import "github.com/google/uuid"
 
 // responseStatus contains the status info of the response.
 type responseStatus struct {
-	code       uint16
+	code       uint32
 	message    string
 	attributes map[string]interface{}
 }
